refactor(brands): extract brand kit services
