+++ apps/server/api/src/collections/brands/brands.module.ts
@@ -9,6 +9,8 @@ import { BrandsController } from '@api/collections/brands/controllers/brands.con
 import { BrandsRelationshipsController } from '@api/collections/brands/controllers/relationships/brands-relationships.controller';
 import { BrandDataMapper } from '@api/collections/brands/services/brand-data.mapper';
 import { BrandGenerationService } from '@api/collections/brands/services/brand-generation.service';
+import { BrandKitAssetsService } from '@api/collections/brands/services/brand-kit-assets.service';
+import { BrandKitDraftService } from '@api/collections/brands/services/brand-kit-draft.service';
 import { BrandPersistenceService } from '@api/collections/brands/services/brand-persistence.service';
 import { BrandRelocationService } from '@api/collections/brands/services/brand-relocation.service';
 import { BrandSetupService } from '@api/collections/brands/services/brand-setup.service';
@@ -72,6 +74,8 @@ import { forwardRef, Module } from '@nestjs/common';
     // OnboardingModule ↔ BrandsModule import cycle.
     BrandSetupService,
     BrandGenerationService,
+    BrandKitAssetsService,
+    BrandKitDraftService,
     BrandPersistenceService,
     BrandRelocationService,
     BrandDataMapper,
