+++ apps/server/api/src/collections/brands/services/brand-kit-assets.service.ts
@@ -1,0 +1,479 @@
+import type { BrandDocument } from '@api/collections/brands/schemas/brand.schema';
+import {
+  CACHE_PATTERNS,
+  CACHE_TAGS,
+} from '@api/common/constants/cache-patterns.constants';
+import { CacheInvalidationService } from '@api/common/services/cache-invalidation.service';
+import { NotFoundException } from '@api/helpers/exceptions/http/not-found.exception';
+import { assertUrlNotPrivate } from '@api/helpers/utils/ssrf/ssrf.util';
+import { PrismaService } from '@api/shared/modules/prisma/prisma.service';
+import { FileInputType } from '@genfeedai/enums';
+import type {
+  BrandKitAssetRole,
+  IBrandKitAssetImportCandidate,
+  IBrandKitAssetImportRequest,
+  IBrandKitAssetImportResponse,
+  IBrandKitAssetImportResult,
+  IBrandKitDiagnostic,
+} from '@genfeedai/interfaces';
+import { Prisma } from '@genfeedai/prisma';
+import { Injectable } from '@nestjs/common';
+import { FilesClientService } from '@server/services/files-microservice/client/files-client.service';
+
+const BRAND_KIT_IMPORT_MAX_BYTES = 50 * 1024 * 1024;
+const BRAND_KIT_ALLOWED_MIME_TYPES = new Set([
+  'image/gif',
+  'image/jpeg',
+  'image/jpg',
+  'image/png',
+  'image/webp',
+]);
+const BRAND_KIT_ALLOWED_EXTENSIONS = new Set([
+  '.gif',
+  '.jpeg',
+  '.jpg',
+  '.png',
+  '.webp',
+]);
+const PRISMA_ASSET_CATEGORY_BY_ROLE: Record<
+  BrandKitAssetRole,
+  Prisma.AssetCreateInput['category']
+> = {
+  banner: 'BANNER' as Prisma.AssetCreateInput['category'],
+  logo: 'LOGO' as Prisma.AssetCreateInput['category'],
+  reference: 'REFERENCE' as Prisma.AssetCreateInput['category'],
+};
+const ASSET_UPLOAD_TYPE_BY_ROLE: Record<BrandKitAssetRole, string> = {
+  banner: 'banners',
+  logo: 'logos',
+  reference: 'references',
+};
+
+type BrandKitAssetBrandFinder = (
+  criteria: Record<string, unknown>,
+) => Promise<BrandDocument | null>;
+
+@Injectable()
+export class BrandKitAssetsService {
+  constructor(
+    private readonly prisma: PrismaService,
+    private readonly cacheInvalidationService: CacheInvalidationService,
+    private readonly filesClientService: FilesClientService,
+  ) {}
+
+  async importBrandKitAssets(
+    brandId: string,
+    organizationId: string,
+    userId: string,
+    dto: IBrandKitAssetImportRequest,
+    findBrand: BrandKitAssetBrandFinder,
+  ): Promise<IBrandKitAssetImportResponse> {
+    const brand = await findBrand({
+      id: brandId,
+      isDeleted: false,
+      organizationId,
+    });
+
+    if (!brand) {
+      throw new NotFoundException('Brand', brandId);
+    }
+
+    if (dto.assets.length === 0) {
+      const diagnostic = this.createBrandKitImportDiagnostic(
+        'brand_kit_asset_import_empty',
+        'At least one asset candidate is required.',
+        'error',
+      );
+
+      return {
+        brandId,
+        diagnostics: [diagnostic],
+        failedCandidateIds: [],
+        importedAssetIds: [],
+        results: [],
+        skippedCandidateIds: [],
+        status: 'blocked',
+      };
+    }
+
+    const results: IBrandKitAssetImportResult[] = [];
+    for (const candidate of dto.assets) {
+      results.push(
+        await this.importBrandKitAssetCandidate(
+          candidate,
+          brandId,
+          organizationId,
+          userId,
+        ),
+      );
+    }
+
+    const importedAssetIds = results
+      .filter((result) => result.status === 'imported' && result.assetId)
+      .map((result) => String(result.assetId));
+    const skippedCandidateIds = results
+      .filter((result) => result.status === 'skipped' && result.candidateId)
+      .map((result) => String(result.candidateId));
+    const failedCandidateIds = results
+      .filter((result) => result.status === 'failed' && result.candidateId)
+      .map((result) => String(result.candidateId));
+    const diagnostics = results.flatMap((result) => result.diagnostics);
+
+    if (importedAssetIds.length > 0) {
+      await this.invalidateBrandAssetCaches(brandId, organizationId);
+    }
+
+    return {
+      brandId,
+      diagnostics,
+      failedCandidateIds,
+      importedAssetIds,
+      results,
+      skippedCandidateIds,
+      status:
+        importedAssetIds.length === results.length
+          ? 'accepted'
+          : importedAssetIds.length > 0
+            ? 'partial'
+            : 'blocked',
+    };
+  }
+
+  private async importBrandKitAssetCandidate(
+    candidate: IBrandKitAssetImportCandidate,
+    brandId: string,
+    organizationId: string,
+    userId: string,
+  ): Promise<IBrandKitAssetImportResult> {
+    const candidateId = candidate.candidateId;
+    const validation = this.validateBrandKitAssetCandidate(candidate);
+
+    if (validation.diagnostics.length > 0) {
+      return {
+        candidateId,
+        diagnostics: validation.diagnostics,
+        role: candidate.role,
+        status: 'failed',
+      };
+    }
+
+    const sourceUrl = validation.url.href;
+    const category = PRISMA_ASSET_CATEGORY_BY_ROLE[candidate.role];
+    const existing = await this.prisma.asset.findFirst({
+      where: {
+        category,
+        isDeleted: false,
+        origin: sourceUrl,
+        parentBrandId: brandId,
+        parentOrgId: organizationId,
+        parentType: 'BRAND' as Prisma.AssetCreateInput['parentType'],
+      },
+    });
+
+    if (existing) {
+      return {
+        assetId: existing.id,
+        candidateId,
+        diagnostics: [
+          this.createBrandKitImportDiagnostic(
+            'brand_kit_asset_already_imported',
+            `${candidate.role} candidate was already imported.`,
+            'info',
+          ),
+        ],
+        role: candidate.role,
+        status: 'skipped',
+        url: this.buildImportedAssetUrl(existing.id, candidate.role),
+      };
+    }
+
+    const hasExistingPrimary =
+      candidate.role !== 'reference'
+        ? await this.hasExistingBrandAsset(
+            brandId,
+            organizationId,
+            candidate.role,
+          )
+        : false;
+
+    if (hasExistingPrimary && !candidate.replaceExisting) {
+      return {
+        candidateId,
+        diagnostics: [
+          this.createBrandKitImportDiagnostic(
+            'brand_kit_asset_existing_preserved',
+            `Existing brand ${candidate.role} was preserved. Set replaceExisting to import this candidate.`,
+            'warning',
+          ),
+        ],
+        role: candidate.role,
+        status: 'skipped',
+      };
+    }
+
+    const asset = await this.prisma.asset.create({
+      data: {
+        category,
+        displayName: candidate.label,
+        mimeType: validation.mimeType,
+        origin: sourceUrl,
+        originalFileName: this.readFileName(validation.url),
+        parentBrandId: brandId,
+        parentOrgId: organizationId,
+        parentType: 'BRAND' as Prisma.AssetCreateInput['parentType'],
+        residency: 'cloud',
+        uploadPolicy: 'brand_kit_import',
+        userId,
+      } satisfies Prisma.AssetUncheckedCreateInput,
+    });
+
+    try {
+      const uploadMeta = await this.filesClientService.uploadToS3(
+        asset.id,
+        ASSET_UPLOAD_TYPE_BY_ROLE[candidate.role],
+        {
+          type: FileInputType.URL,
+          url: sourceUrl,
+        },
+      );
+
+      if (
+        typeof uploadMeta.size === 'number' &&
+        uploadMeta.size > BRAND_KIT_IMPORT_MAX_BYTES
+      ) {
+        await this.markImportedAssetDeleted(asset.id);
+        return {
+          candidateId,
+          diagnostics: [
+            this.createBrandKitImportDiagnostic(
+              'brand_kit_asset_too_large',
+              `Imported asset exceeds the ${BRAND_KIT_IMPORT_MAX_BYTES / (1024 * 1024)}MB brand kit limit.`,
+              'error',
+            ),
+          ],
+          role: candidate.role,
+          status: 'failed',
+        };
+      }
+
+      const publicUrl =
+        typeof uploadMeta.publicUrl === 'string'
+          ? uploadMeta.publicUrl
+          : this.buildImportedAssetUrl(asset.id, candidate.role);
+
+      await this.prisma.asset.update({
+        where: { id: asset.id },
+        data: {
+          cloudObjectKey: `ingredients/${ASSET_UPLOAD_TYPE_BY_ROLE[candidate.role]}/${asset.id}`,
+          mimeType: validation.mimeType,
+          sizeBytes:
+            typeof uploadMeta.size === 'number' ? uploadMeta.size : undefined,
+        },
+      });
+
+      if (candidate.replaceExisting) {
+        await this.softDeleteReplacedBrandAssets(
+          brandId,
+          organizationId,
+          candidate.role,
+          asset.id,
+        );
+      }
+
+      return {
+        assetId: asset.id,
+        candidateId,
+        diagnostics: [],
+        role: candidate.role,
+        status: 'imported',
+        url: publicUrl,
+      };
+    } catch (error: unknown) {
+      await this.markImportedAssetDeleted(asset.id);
+      const message =
+        error instanceof Error ? error.message : 'Remote asset import failed';
+      return {
+        candidateId,
+        diagnostics: [
+          this.createBrandKitImportDiagnostic(
+            'brand_kit_asset_import_failed',
+            message,
+            'error',
+          ),
+        ],
+        role: candidate.role,
+        status: 'failed',
+      };
+    }
+  }
+
+  private validateBrandKitAssetCandidate(
+    candidate: IBrandKitAssetImportCandidate,
+  ): {
+    diagnostics: IBrandKitDiagnostic[];
+    mimeType?: string;
+    url: URL;
+  } {
+    const diagnostics: IBrandKitDiagnostic[] = [];
+    const rawUrl = candidate.url ?? candidate.sourceUrl;
+    let parsedUrl: URL | undefined;
+
+    if (!rawUrl) {
+      diagnostics.push(
+        this.createBrandKitImportDiagnostic(
+          'brand_kit_asset_missing_url',
+          'Asset candidate requires a URL.',
+          'error',
+        ),
+      );
+    } else {
+      try {
+        parsedUrl = new URL(rawUrl);
+        if (!['http:', 'https:'].includes(parsedUrl.protocol)) {
+          diagnostics.push(
+            this.createBrandKitImportDiagnostic(
+              'brand_kit_asset_invalid_protocol',
+              'Asset candidate URL must use http or https.',
+              'error',
+            ),
+          );
+        } else {
+          assertUrlNotPrivate(parsedUrl.href);
+        }
+      } catch (error: unknown) {
+        const message =
+          error instanceof Error
+            ? error.message
+            : 'Invalid asset candidate URL';
+        diagnostics.push(
+          this.createBrandKitImportDiagnostic(
+            'brand_kit_asset_invalid_url',
+            message,
+            'error',
+          ),
+        );
+      }
+    }
+
+    const normalizedMimeType = candidate.mimeType?.trim().toLowerCase();
+    if (
+      normalizedMimeType &&
+      !BRAND_KIT_ALLOWED_MIME_TYPES.has(normalizedMimeType)
+    ) {
+      diagnostics.push(
+        this.createBrandKitImportDiagnostic(
+          'brand_kit_asset_unsupported_content_type',
+          `${normalizedMimeType} is not supported for brand kit assets.`,
+          'error',
+        ),
+      );
+    }
+
+    if (!normalizedMimeType && parsedUrl) {
+      const extension = this.readExtension(parsedUrl);
+      if (!BRAND_KIT_ALLOWED_EXTENSIONS.has(extension)) {
+        diagnostics.push(
+          this.createBrandKitImportDiagnostic(
+            'brand_kit_asset_unknown_content_type',
+            'Asset candidate must include a supported image MIME type or file extension.',
+            'error',
+          ),
+        );
+      }
+    }
+
+    return {
+      diagnostics,
+      mimeType: normalizedMimeType,
+      url: parsedUrl ?? new URL('https://invalid.example'),
+    };
+  }
+
+  private async hasExistingBrandAsset(
+    brandId: string,
+    organizationId: string,
+    role: Exclude<BrandKitAssetRole, 'reference'>,
+  ): Promise<boolean> {
+    const existing = await this.prisma.asset.findFirst({
+      select: { id: true },
+      where: {
+        category: PRISMA_ASSET_CATEGORY_BY_ROLE[role],
+        isDeleted: false,
+        parentBrandId: brandId,
+        parentOrgId: organizationId,
+        parentType: 'BRAND' as Prisma.AssetCreateInput['parentType'],
+      },
+    });
+
+    return Boolean(existing);
+  }
+
+  private async softDeleteReplacedBrandAssets(
+    brandId: string,
+    organizationId: string,
+    role: BrandKitAssetRole,
+    keepAssetId: string,
+  ): Promise<void> {
+    await this.prisma.asset.updateMany({
+      where: {
+        category: PRISMA_ASSET_CATEGORY_BY_ROLE[role],
+        id: { not: keepAssetId },
+        isDeleted: false,
+        parentBrandId: brandId,
+        parentOrgId: organizationId,
+        parentType: 'BRAND' as Prisma.AssetCreateInput['parentType'],
+      },
+      data: { isDeleted: true },
+    });
+  }
+
+  private async markImportedAssetDeleted(assetId: string): Promise<void> {
+    await this.prisma.asset.update({
+      where: { id: assetId },
+      data: { isDeleted: true },
+    });
+  }
+
+  private async invalidateBrandAssetCaches(
+    brandId: string,
+    organizationId: string,
+  ): Promise<void> {
+    await this.cacheInvalidationService.invalidate(
+      CACHE_PATTERNS.BRANDS_SINGLE(brandId),
+      CACHE_PATTERNS.BRANDS_LIST(organizationId),
+      `brand:${brandId}`,
+    );
+    await this.cacheInvalidationService.invalidateByTags([
+      CACHE_TAGS.BRANDS,
+      'assets',
+      'links',
+      'public',
+    ]);
+  }
+
+  private createBrandKitImportDiagnostic(
+    code: string,
+    message: string,
+    severity: IBrandKitDiagnostic['severity'],
+  ): IBrandKitDiagnostic {
+    return { code, message, severity };
+  }
+
+  private buildImportedAssetUrl(
+    assetId: string,
+    role: BrandKitAssetRole,
+  ): string {
+    return `/ingredients/${ASSET_UPLOAD_TYPE_BY_ROLE[role]}/${assetId}`;
+  }
+
+  private readExtension(url: URL): string {
+    const pathname = url.pathname.toLowerCase();
+    const extensionStart = pathname.lastIndexOf('.');
+    return extensionStart >= 0 ? pathname.slice(extensionStart) : '';
+  }
+
+  private readFileName(url: URL): string | undefined {
+    const filename = url.pathname.split('/').filter(Boolean).at(-1);
+    return filename ? decodeURIComponent(filename).slice(0, 180) : undefined;
+  }
+}
